Add paging to SortAlgo, change uint16_t to unsigned int

+++ DMA/ipos-gcc-dma/include/pTCB.h
@@ -14,8 +14,8 @@
 void os_memMapper(unsigned int *cnt, taskId _task);
 
 void os_addTasks(unsigned char numTasks, taskId tasks[]);
-uint16_t* os_search(funcPt func);
+unsigned int* os_search(funcPt func);
 
-void __os_block(funcPt func[], uint16_t numTasks);
-void __os_unblock(funcPt func[], uint16_t numTasks);
+void __os_block(funcPt func[], unsigned int numTasks);
+void __os_unblock(funcPt func[], unsigned int numTasks);
 #endif /* INCLUDE_PERSISTENTLIST_H_ */
